API Documentation

Exported

Orthography.codepoints — Function

Delegate to specific functions based on type's orthography trait value.

codepoints(x)

It is an error to invoke the codepoints function on anything but an orthographic system.

codepoints(_, x)

Orthographic systems must implement codepoints.

codepoints(_, ortho)

Implement codepoints function for SimpleAscii.

codepoints(ortho)

Implement codepoints function for SimpleAscii.

codepoints(ortho)

Implement Orthography's codepoints functions for LiteraryGreekOrthography.

codepoints(ortho)

Implement Orthography's codepoints function.

codepoints(ortho)

Orthography.tokentypes — Function

Delegate to specific functions based on type's orthography trait value.

tokentypes(x)

It is an error to invoke the tokentypes function on anything but an orthographic system.

tokentypes(_, x)

Orthographic systems must implement tokentypes.

tokentypes(_, ortho, s)

Implement tokentypes function for SimpleAscii.

tokentypes(ortho)

Implement tokentypes function for WSTokenizer.

tokentypes(ortho)

Implement Orthography's tokentypes functions for LiteraryGreekOrthography.

tokentypes(ortho)

Implement Orthography's tokentypes function for AtticOrthography.

tokentypes(ortho)

Orthography.tokenize — Function

Delegate to specific functions based on type's orthography trait value.

tokenize(s, x)

It is an error to invoke the tokenize function on anything but an orthographic system.

tokenize(_, s, x)

Orthographic systems must implement tokenize.

tokenize(_, s, ortho)

Tokenize citable node cn using the tokenizer of the given orthographic system.

tokenize(psg, ortho; edition, exemplar)

The return value is a list of pairings of a CitablePassage and a token category. The citable node is citable at the level of the token.

Tokenize corpus c using the tokenizer of the given orthographic system.

tokenize(c, ortho; edition, exemplar)

The return value is a list of pairings of a CitablePassage and a token category. The citable node is citable at the level of the token.

Tokenize document doc using the tokenizer of the given orthographic system.

tokenize(doc, ortho; edition, exemplar)

The return value is a list of pairings of a CitablePassage and a token category. The citable node is citable at the level of the token.

Implement tokenize function for SimpleAscii orthography.

tokenize(s, o)

Implement tokenize function for WSTokenizer orthography.

tokenize(s, o)

Implement Orthography's tokenize function for LiteraryGreekOrthography.

tokenize(s, o)

Tokenize a string in orthography of HMT Greek MSS.

tokenize(s, o)

Internals

ManuscriptOrthography.splitPunctuation — Function

Split off any trailing punctuation and return an Array of leading string + trailing punctuation.

splitPunctuation(s)